separate endpoints from database functions
