+++ api/database/__init__.py
@@ -1,1 +1,7 @@
-"""Test"""
+"""Database functions."""
+from pymongo import MongoClient
+
+
+def create_db_client(uri):
+    """Connect to database and return db client."""
+    return MongoClient(uri)
