+++ api/endpoints/__init__.py
@@ -1,0 +1,1 @@
+"""API endpoint processors."""
